requirements.txt
absl-py
pytorch_lightning
nlp
transformers
sh